refactor(workload): split registry internals
